Restructure baseapp
