--- baseapp/doc.go
@@ -1,8 +1,0 @@
-/*
-Package baseapp contains data structures that provide basic data storage
-functionality and act as a bridge between the ABCI interface and the SDK
-abstractions.
-
-BaseApp has no state except the CommitMultiStore you provide upon init.
-*/
-package baseapp
